closing the modal via the ESC(escape key)
The <dialog> element allows website visitors to close the opened dialog by pressing the ESC (Escape) key on their keyboard.

Currently, this will not trigger the onReset function though (unlike closing the dialog with a button click).

To make sure that onReset gets triggered when the dialog is closed via the escape key, you should add the built-in onClose prop to the <dialog> element and bind it to the onReset prop value.

Like this:

<dialog ref={dialog} className="result-modal" onClose={onReset}
   ...
</dialog>